Refactor rich text editor and its dependencies to their respective files
