+++ src/lib/types/iv-editor.ts
@@ -1,0 +1,2 @@
+export type CustomElement = { type: 'paragraph'; children: CustomText[] };
+export type CustomText = { text: string };
